# frontend/src/lib/api/swagger3.json
{
    "openapi": "3.0.0",
...
    "info": {
        "description": "API системы мониторинга репутации брендов BrandRadar",
        "title": "BrandRadar API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/mentions": {
            "get": {
                "description": "Возвращает упоминания с фильтрацией и пагинацией",
                "tags": [
                    "Mentions"
                ],
                "summary": "Список упоминаний",
                "parameters": [
                    {
                        "description": "UUID бренда",
                        "name": "brand_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "UUID источника",
                        "name": "source_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Статус (new, processing, processed, archived, discarded)",
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ML-метка (negative, positive, neutral)",
                        "name": "label",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Релевантность",
                        "name": "is_relevant",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "Поиск по тексту",
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Дата от (RFC3339)",
                        "name": "date_from",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Дата до (RFC3339)",
                        "name": "date_to",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "allOf": [
                                                        {
                                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                                                        },
                                                        {
                                                            "type": "object",
                                                            "properties": {
                                                                "items": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.MentionResponse"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    ]
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/mentions/{id}": {
            "get": {
                "description": "Возвращает упоминание с ML-полями по ID",
                "tags": [
                    "Mentions"
                ],
                "summary": "Получить упоминание",
                "parameters": [
                    {
                        "description": "UUID упоминания",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.MentionResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "NOT_FOUND"
                    },
                    "message": {
                        "type": "string",
                        "example": "resource not found"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.ErrorResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.ErrorDetail"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.MentionMLResponse": {
                "type": "object",
                "properties": {
                    "is_relevant": {
                        "type": "boolean",
                        "example": true
                    },
                    "label": {
                        "type": "string",
                        "example": "negative"
                    },
                    "score": {
                        "type": "number",
                        "example": 0.87
                    },
                    "similar_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "550e8400-e29b-41d4-a716-446655440000"
                        ]
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.MentionResponse": {
                "type": "object",
                "properties": {
                    "author": {
                        "type": "string",
                        "example": "@user"
                    },
                    "brand_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "deduplicated": {
                        "type": "boolean",
                        "example": false
                    },
                    "external_id": {
                        "type": "string",
                        "example": "tg-12345"
                    },
                    "id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "ml": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.MentionMLResponse"
                    },
                    "published_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "source_id": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "status": {
                        "type": "string",
                        "example": "new"
                    },
                    "text": {
                        "type": "string",
                        "example": "Текст публикации"
                    },
                    "title": {
                        "type": "string",
                        "example": "Заголовок"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://t.me/channel/123"
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData": {
                "type": "object",
                "properties": {
                    "items": {},
                    "limit": {
                        "type": "integer",
                        "example": 20
                    },
                    "offset": {
                        "type": "integer",
                        "example": 0
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    }
                }
            },
            "prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/prod-pobeda-2026_internal_controller_http_v1_dto.PaginatedData"
                    },
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
...
            "prod-pobeda-2026_internal_controller_http_v1_dto.SuccessResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "type": "string",
                        "example": "null"
                    }
                }
            },
...
        }
    }
}